+++ src/OpenTelemetry/Internal/OpenTelemetrySdkEventSource.cs
@@ -15,6 +15,7 @@
 // </copyright>
 
 #nullable enable
+
 using System.Diagnostics;
 #if NET6_0_OR_GREATER
 using System.Diagnostics.CodeAnalysis;