+++ src/OpenTelemetry/Metrics/ExponentialHistogramData.cs
@@ -18,6 +18,9 @@
 
 namespace OpenTelemetry.Metrics;
 
+/// <summary>
+/// Contains the data for an exponential histogram.
+/// </summary>
 public sealed class ExponentialHistogramData
 {
     internal ExponentialHistogramData()
@@ -26,21 +29,35 @@ public sealed class ExponentialHistogramData
         this.NegativeBuckets = new();
     }
 
+    /// <summary>
+    /// Gets the exponential histogram scale.
+    /// </summary>
     public int Scale { get; internal set; }
 
+    /// <summary>
+    /// Gets the exponential histogram zero count.
+    /// </summary>
     public long ZeroCount { get; internal set; }
 
+    /// <summary>
+    /// Gets the exponential histogram positive buckets.
+    /// </summary>
     public ExponentialHistogramBuckets PositiveBuckets { get; private set; }
 
+    /// <summary>
+    /// Gets the exponential histogram negative buckets.
+    /// </summary>
     internal ExponentialHistogramBuckets NegativeBuckets { get; private set; }
 
     internal ExponentialHistogramData Copy()
     {
-        var copy = new ExponentialHistogramData();
-        copy.Scale = this.Scale;
-        copy.ZeroCount = this.ZeroCount;
-        copy.PositiveBuckets = this.PositiveBuckets.Copy();
-        copy.NegativeBuckets = this.NegativeBuckets.Copy();
+        var copy = new ExponentialHistogramData
+        {
+            Scale = this.Scale,
+            ZeroCount = this.ZeroCount,
+            PositiveBuckets = this.PositiveBuckets.Copy(),
+            NegativeBuckets = this.NegativeBuckets.Copy(),
+        };
         return copy;
     }
 }