+++ src/OpenTelemetry/Metrics/MetricPoint.cs
@@ -45,7 +45,7 @@ public struct MetricPoint
     internal MetricPoint(
         AggregatorStore aggregatorStore,
         AggregationType aggType,
-        KeyValuePair<string, object>[] tagKeysAndValues,
+        KeyValuePair<string, object?>[] tagKeysAndValues,
         double[] histogramExplicitBounds,
         int exponentialHistogramMaxSize,
         int exponentialHistogramMaxScale)
@@ -304,7 +304,7 @@ public struct MetricPoint
     /// <param name="max"> The histogram maximum value.</param>
     /// <returns>True if minimum and maximum value exist, false otherwise.</returns>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public bool TryGetHistogramMinMaxValues(out double min, out double max)
+    public readonly bool TryGetHistogramMinMaxValues(out double min, out double max)
     {
         if (this.aggType == AggregationType.HistogramWithMinMax ||
                         this.aggType == AggregationType.HistogramWithMinMaxBuckets)
@@ -432,7 +432,7 @@ public struct MetricPoint
         this.MetricPointStatus = MetricPointStatus.CollectPending;
     }
 
-    internal void UpdateWithExemplar(long number, ReadOnlySpan<KeyValuePair<string, object>> tags, bool isSampled)
+    internal void UpdateWithExemplar(long number, ReadOnlySpan<KeyValuePair<string, object?>> tags, bool isSampled)
     {
         switch (this.aggType)
         {
@@ -629,7 +629,7 @@ public struct MetricPoint
         this.MetricPointStatus = MetricPointStatus.CollectPending;
     }
 
-    internal void UpdateWithExemplar(double number, ReadOnlySpan<KeyValuePair<string, object>> tags, bool isSampled)
+    internal void UpdateWithExemplar(double number, ReadOnlySpan<KeyValuePair<string, object?>> tags, bool isSampled)
     {
         switch (this.aggType)
         {
@@ -855,7 +855,9 @@ public struct MetricPoint
                         histogramBuckets.RunningSum = 0;
                     }
 
-                    for (int i = 0; i < histogramBuckets.RunningBucketCounts.Length; i++)
+                    Debug.Assert(histogramBuckets.RunningBucketCounts != null, "histogramBuckets.RunningBucketCounts was null");
+
+                    for (int i = 0; i < histogramBuckets.RunningBucketCounts!.Length; i++)
                     {
                         histogramBuckets.SnapshotBucketCounts[i] = histogramBuckets.RunningBucketCounts[i];
                         if (outputDelta)
@@ -913,7 +915,9 @@ public struct MetricPoint
                         histogramBuckets.RunningMax = double.NegativeInfinity;
                     }
 
-                    for (int i = 0; i < histogramBuckets.RunningBucketCounts.Length; i++)
+                    Debug.Assert(histogramBuckets.RunningBucketCounts != null, "histogramBuckets.RunningBucketCounts was null");
+
+                    for (int i = 0; i < histogramBuckets.RunningBucketCounts!.Length; i++)
                     {
                         histogramBuckets.SnapshotBucketCounts[i] = histogramBuckets.RunningBucketCounts[i];
                         if (outputDelta)
@@ -1103,7 +1107,9 @@ public struct MetricPoint
                         histogramBuckets.RunningSum = 0;
                     }
 
-                    for (int i = 0; i < histogramBuckets.RunningBucketCounts.Length; i++)
+                    Debug.Assert(histogramBuckets.RunningBucketCounts != null, "histogramBuckets.RunningBucketCounts was null");
+
+                    for (int i = 0; i < histogramBuckets.RunningBucketCounts!.Length; i++)
                     {
                         histogramBuckets.SnapshotBucketCounts[i] = histogramBuckets.RunningBucketCounts[i];
                         if (outputDelta)
@@ -1163,7 +1169,9 @@ public struct MetricPoint
                         histogramBuckets.RunningMax = double.NegativeInfinity;
                     }
 
-                    for (int i = 0; i < histogramBuckets.RunningBucketCounts.Length; i++)
+                    Debug.Assert(histogramBuckets.RunningBucketCounts != null, "histogramBuckets.RunningBucketCounts was null");
+
+                    for (int i = 0; i < histogramBuckets.RunningBucketCounts!.Length; i++)
                     {
                         histogramBuckets.SnapshotBucketCounts[i] = histogramBuckets.RunningBucketCounts[i];
                         if (outputDelta)
@@ -1275,7 +1283,7 @@ public struct MetricPoint
         Interlocked.Exchange(ref isCriticalSectionOccupied, 0);
     }
 
-    private void UpdateHistogram(double number, ReadOnlySpan<KeyValuePair<string, object>> tags = default, bool reportExemplar = false)
+    private void UpdateHistogram(double number, ReadOnlySpan<KeyValuePair<string, object?>> tags = default, bool reportExemplar = false)
     {
         var histogramBuckets = this.mpComponents!.HistogramBuckets;
 
@@ -1297,7 +1305,7 @@ public struct MetricPoint
         ReleaseLock(ref histogramBuckets.IsCriticalSectionOccupied);
     }
 
-    private void UpdateHistogramWithMinMax(double number, ReadOnlySpan<KeyValuePair<string, object>> tags = default, bool reportExemplar = false)
+    private void UpdateHistogramWithMinMax(double number, ReadOnlySpan<KeyValuePair<string, object?>> tags = default, bool reportExemplar = false)
     {
         var histogramBuckets = this.mpComponents!.HistogramBuckets;
 
@@ -1321,10 +1329,12 @@ public struct MetricPoint
         ReleaseLock(ref histogramBuckets.IsCriticalSectionOccupied);
     }
 
-    private void UpdateHistogramWithBuckets(double number, ReadOnlySpan<KeyValuePair<string, object>> tags = default, bool reportExemplar = false)
+    private void UpdateHistogramWithBuckets(double number, ReadOnlySpan<KeyValuePair<string, object?>> tags = default, bool reportExemplar = false)
     {
         var histogramBuckets = this.mpComponents!.HistogramBuckets;
         int i = histogramBuckets.FindBucketIndex(number);
+
+        Debug.Assert(histogramBuckets.RunningBucketCounts != null, "histogramBuckets.RunningBucketCounts was null");
 
         AcquireLock(ref histogramBuckets.IsCriticalSectionOccupied);
 
@@ -1332,7 +1342,7 @@ public struct MetricPoint
         {
             this.runningValue.AsLong++;
             histogramBuckets.RunningSum += number;
-            histogramBuckets.RunningBucketCounts[i]++;
+            histogramBuckets.RunningBucketCounts![i]++;
             if (reportExemplar)
             {
                 // TODO: Need to ensure that the lock is always released.
@@ -1344,18 +1354,20 @@ public struct MetricPoint
         ReleaseLock(ref histogramBuckets.IsCriticalSectionOccupied);
     }
 
-    private void UpdateHistogramWithBucketsAndMinMax(double number, ReadOnlySpan<KeyValuePair<string, object>> tags = default, bool reportExemplar = false)
+    private void UpdateHistogramWithBucketsAndMinMax(double number, ReadOnlySpan<KeyValuePair<string, object?>> tags = default, bool reportExemplar = false)
     {
         var histogramBuckets = this.mpComponents!.HistogramBuckets;
         int i = histogramBuckets.FindBucketIndex(number);
 
         AcquireLock(ref histogramBuckets.IsCriticalSectionOccupied);
 
+        Debug.Assert(histogramBuckets.RunningBucketCounts != null, "histogramBuckets.RunningBucketCounts was null");
+
         unchecked
         {
             this.runningValue.AsLong++;
             histogramBuckets.RunningSum += number;
-            histogramBuckets.RunningBucketCounts[i]++;
+            histogramBuckets.RunningBucketCounts![i]++;
             if (reportExemplar)
             {
                 // TODO: Need to ensure that the lock is always released.
@@ -1371,7 +1383,7 @@ public struct MetricPoint
     }
 
 #pragma warning disable IDE0060 // Remove unused parameter: Exemplars for exponential histograms will be a follow up PR
-    private void UpdateBase2ExponentialHistogram(double number, ReadOnlySpan<KeyValuePair<string, object>> tags = default, bool reportExemplar = false)
+    private void UpdateBase2ExponentialHistogram(double number, ReadOnlySpan<KeyValuePair<string, object?>> tags = default, bool reportExemplar = false)
 #pragma warning restore IDE0060 // Remove unused parameter
     {
         if (number < 0)
@@ -1394,7 +1406,7 @@ public struct MetricPoint
     }
 
 #pragma warning disable IDE0060 // Remove unused parameter: Exemplars for exponential histograms will be a follow up PR
-    private void UpdateBase2ExponentialHistogramWithMinMax(double number, ReadOnlySpan<KeyValuePair<string, object>> tags = default, bool reportExemplar = false)
+    private void UpdateBase2ExponentialHistogramWithMinMax(double number, ReadOnlySpan<KeyValuePair<string, object?>> tags = default, bool reportExemplar = false)
 #pragma warning restore IDE0060 // Remove unused parameter
     {
         if (number < 0)