+++ src/OpenTelemetry/Metrics/PeriodicExportingMetricReaderOptions.cs
@@ -14,11 +14,20 @@
 // limitations under the License.
 // </copyright>
 
+#nullable enable
+
 using Microsoft.Extensions.Configuration;
 using OpenTelemetry.Internal;
 
 namespace OpenTelemetry.Metrics;
 
+/// <summary>
+/// Contains periodic metric reader options.
+/// </summary>
+/// <remarks>
+/// Note: OTEL_METRIC_EXPORT_INTERVAL and OTEL_METRIC_EXPORT_TIMEOUT environment
+/// variables are parsed during object construction.
+/// </remarks>
 public class PeriodicExportingMetricReaderOptions
 {
     internal const string OTelMetricExportIntervalEnvVarKey = "OTEL_METRIC_EXPORT_INTERVAL";