+++ src/OpenTelemetry/Trace/Sampler.cs
@@ -25,6 +25,9 @@ namespace OpenTelemetry.Trace;
 /// </summary>
 public abstract class Sampler
 {
+    /// <summary>
+    /// Initializes a new instance of the <see cref="Sampler"/> class.
+    /// </summary>
     protected Sampler()
     {
         this.Description = this.GetType().Name;