+++ src/OpenTelemetry.Shims.OpenTracing/CHANGELOG.md
@@ -2,9 +2,17 @@
 
 ## Unreleased
 
+## 1.6.0-beta.1
+
+Released 2023-Sep-05
+
 * Fix: Do not raise `ArgumentException` if `Activity` behind the shim span
   has an invalid context.
   ([#2787](https://github.com/open-telemetry/opentelemetry-dotnet/issues/2787))
+* Obsolete `TracerShim(Tracer, TextMapPropagator)` constructor.
+  Provide `TracerShim(TracerProvider)`
+  and `TracerShim(TracerProvider, TextMapPropagator)` constructors.
+  ([#4812](https://github.com/open-telemetry/opentelemetry-dotnet/pull/4812))
 
 ## 1.5.0-beta.1